+++ common/src/main/java/com/avp/common/constant/ArmorConstants.java
@@ -6,6 +6,8 @@ public record ArmorConstants() {
 
     public static final float PRAETORIAN_ARMOR = 12.0F;
 
+    public static final float PROWLER_ARMOR = 8.0F;
+
     public static final float QUEEN_ARMOR = 16.0F;
 
     public static final float RUNNER_ARMOR = 4.0F;