+++ common/src/main/java/com/avp/common/constant/AttackDamageConstants.java
@@ -14,6 +14,8 @@ public record AttackDamageConstants() {
 
     public static final float PRAETORIAN_ATTACK_DAMAGE = PLAYER_HEALTH * 0.75F;
 
+    public static final float PROWLER_ATTACK_DAMAGE = PLAYER_HEALTH * 0.5F;
+
     public static final float QUEEN_ATTACK_DAMAGE = PLAYER_HEALTH * 2.5F;
 
     public static final float RUNNER_ATTACK_DAMAGE = PLAYER_HEALTH * 0.25F;