+++ common/src/main/java/com/avp/common/constant/FollowRangeConstants.java
@@ -10,6 +10,8 @@ public record FollowRangeConstants() {
 
     public static final float PRAETORIAN_FOLLOW_RANGE = 35F;
 
+    public static final float PROWLER_FOLLOW_RANGE = 35F;
+
     public static final float QUEEN_FOLLOW_RANGE = 35F;
 
     public static final float RUNNER_FOLLOW_RANGE = 35F;