+++ fabric/src/main/java/com/avp/fabric/data/lang/en_us/provider/EnUsEntityProvider.java
@@ -23,6 +23,7 @@ public class EnUsEntityProvider {
         addEntity(builder, AlienEntityTypes.ABERRANT_FACEHUGGER, "Aberrant Facehugger");
         addEntity(builder, AlienEntityTypes.ABERRANT_OVOMORPH, "Aberrant Ovomorph");
         addEntity(builder, AlienEntityTypes.ABERRANT_PRAETORIAN, "Aberrant Praetorian");
+        addEntity(builder, AlienEntityTypes.ABERRANT_PROWLER, "Aberrant Prowler");
         addEntity(builder, AlienEntityTypes.ABERRANT_QUEEN, "Aberrant Queen");
         addEntity(builder, AlienEntityTypes.ABERRANT_RUNNER, "Aberrant Runner");
         addEntity(builder, AlienEntityTypes.ABERRANT_WARRIOR, "Aberrant Warrior");
@@ -32,6 +33,7 @@ public class EnUsEntityProvider {
         addEntity(builder, AlienEntityTypes.FACEHUGGER, "Facehugger");
         addEntity(builder, AlienEntityTypes.IRRADIATED_DRONE, "Irradiated Drone");
         addEntity(builder, AlienEntityTypes.IRRADIATED_PRAETORIAN, "Irradiated Praetorian");
+        addEntity(builder, AlienEntityTypes.IRRADIATED_PROWLER, "Irradiated Prowler");
         addEntity(builder, AlienEntityTypes.IRRADIATED_QUEEN, "Irradiated Queen");
         addEntity(builder, AlienEntityTypes.IRRADIATED_RUNNER, "Irradiated Runner");
         addEntity(builder, AlienEntityTypes.IRRADIATED_WARRIOR, "Irradiated Warrior");
@@ -40,11 +42,13 @@ public class EnUsEntityProvider {
         addEntity(builder, AlienEntityTypes.NETHER_FACEHUGGER, "Nether Facehugger");
         addEntity(builder, AlienEntityTypes.NETHER_OVOMORPH, "Nether Ovomorph");
         addEntity(builder, AlienEntityTypes.NETHER_PRAETORIAN, "Nether Praetorian");
+        addEntity(builder, AlienEntityTypes.NETHER_PROWLER, "Nether Prowler");
         addEntity(builder, AlienEntityTypes.NETHER_QUEEN, "Nether Queen");
         addEntity(builder, AlienEntityTypes.NETHER_RUNNER, "Nether Runner");
         addEntity(builder, AlienEntityTypes.NETHER_WARRIOR, "Nether Warrior");
         addEntity(builder, AlienEntityTypes.OVOMORPH, "Ovomorph");
         addEntity(builder, AlienEntityTypes.PRAETORIAN, "Praetorian");
+        addEntity(builder, AlienEntityTypes.PROWLER, "Prowler");
         addEntity(builder, AlienEntityTypes.QUEEN, "Queen");
         addEntity(builder, AlienEntityTypes.ROYAL_ABERRANT_CHESTBURSTER, "Royal Aberrant Chestburster");
         addEntity(builder, AlienEntityTypes.ROYAL_ABERRANT_FACEHUGGER, "Royal Aberrant Facehugger");